+++ Digital Image Processing/src/com/frank/dip/enhance/convolver/HarmonicMeanKernel.java
@@ -149,9 +149,9 @@ public class HarmonicMeanKernel extends SingleKernel
 	}
 
 	/**
-	 * @see com.frank.dip.enhance.convolver.SingleKernel#perform(int[])
+	 * @see com.frank.dip.enhance.convolver.SingleKernel#perform(float[])
 	 */
-	public float perform(int... pixels)
+	public float perform(float... pixels)
 	{
 		if (summary <= 0)
 			return super.perform(pixels);
