+++ Digital Image Processing/src/com/frank/dip/enhance/time/ConvolveEnhance.java
@@ -160,7 +160,7 @@ public class ConvolveEnhance<T extends Image> extends Operator<T, T> implements
 		boolean fh = h % 2 == 0;// the flag whether the height is even
 		int dw = w / 2;
 		int dh = h / 2;
-		int[] pixels = new int[w * h];
+		float[] pixels = new float[w * h];
 		float pixel;
 		GrayImage res = gi.recreate();
 		for (int y = 0; y < height; y++)
@@ -222,7 +222,7 @@ public class ConvolveEnhance<T extends Image> extends Operator<T, T> implements
 		boolean fh = h % 2 == 0;// the flag whether the height is even
 		int dw = w / 2;
 		int dh = h / 2;
-		int[] pixels = new int[w * h];
+		float[] pixels = new float[w * h];
 		float pixel;
 		int[][] gray = new int[height][width];
 		float minGray = Integer.MAX_VALUE;
@@ -292,9 +292,9 @@ public class ConvolveEnhance<T extends Image> extends Operator<T, T> implements
 		int dw = w / 2;
 		int dh = h / 2;
 		// neighbor pixels
-		int[] reds = new int[w * h];
-		int[] greens = new int[w * h];
-		int[] blues = new int[w * h];
+		float[] reds = new float[w * h];
+		float[] greens = new float[w * h];
+		float[] blues = new float[w * h];
 		ColorImage res = ci.recreate();
 		// temporary pixels arrays
 		float pixel;
@@ -388,9 +388,9 @@ public class ConvolveEnhance<T extends Image> extends Operator<T, T> implements
 		int dw = w / 2;
 		int dh = h / 2;
 		// neighbor pixels
-		int[] reds = new int[w * h];
-		int[] greens = new int[w * h];
-		int[] blues = new int[w * h];
+		float[] reds = new float[w * h];
+		float[] greens = new float[w * h];
+		float[] blues = new float[w * h];
 		// temporary pixels arrays
 		int[][] red = new int[height][width];
 		int[][] green = new int[height][width];
